backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.30.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
